requirements.txt
Flask
Werkzeug
Pillow
numpy
pandas
matplotlib
opencv-python
scikit-learn
seaborn
